pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfagent"
version = "0.1.0"
description = "PDF-constrained conversational agent (STAIR x Scaler Task 3)"
requires-python = ">=3.10"
dependencies = [
    "groq>=0.11.0",
    "pymupdf>=1.24.0",
    "rapidocr-onnxruntime>=1.3.0",
    "FlagEmbedding>=1.3.0",
    "chromadb>=0.5.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "streamlit>=1.32.0",
    "pydantic>=2.6.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "langdetect>=1.0.9",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
include = ["pdfagent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
